fix for internal server error behind haproxy
when running the webapp behind haproxy, visits to the page through the
haproxy front-end generated an HTTP 500 error.  looking at the logs,
the following error was present:

192.168.122.1 - - [09/Jan/2023 19:34:59] "GET /video_feed HTTP/1.1" 500 -
Error on request:
Traceback (most recent call last):
  File "/opt/app-root/lib64/python3.9/site-packages/werkzeug/serving.py", line 335, in run_wsgi
    execute(self.server.app)
  File "/opt/app-root/lib64/python3.9/site-packages/werkzeug/serving.py", line 324, in execute
    for data in application_iter:
  File "/opt/app-root/lib64/python3.9/site-packages/werkzeug/wsgi.py", line 500, in __next__
    return self._next()
  File "/opt/app-root/lib64/python3.9/site-packages/werkzeug/wrappers/response.py", line 50, in _iter_encoded
    for item in iterable:
  File "/app/app.py", line 13, in gen
    yield (b'--frame\r\n'
TypeError: can't concat NoneType to bytes

there appeared to be a missing concatenation operator, adding it in
resolved the issue.